Refactor spritesheet into texturearray
